[CodecImage] Verify that CopyTexImage() is functionally a no-op
We are looking to eliminate StreamTextureSharedImageInterface
subclassing GLImage as part of eliminating GLImage as a public
interface. The blocker to doing so is that
SharedImageVideoSurfaceTexture passes a
StreamTextureSharedImageInterface instance to
AbstractTexture::BindStreamTextureImage(), which operates on the
passed-in object as a GLImage. That invocation will in turn result in
the StreamTextureSharedImageInterface instance's BindTexImage() or its
CopyTexImage() method being invoked.

StreamTextureSharedImageInterface has two implementations, and their
implementations of these methods are as follows:

- StreamTexture::BindTexImage(): no-op
- StreamTexture::CopyTexImage(): no-op
- CodecImage::BindTexImage(): no-op
- CodecImage::CopyTexImage(): renders to front buffer

Hence, we need to focus only on CodecImage::CopyTexImage(). Our
suspicion is that it is currently always the case that the image has
already been rendered to the front buffer at this point, making
CodecImage::CopyTexImage() itself a no-op.

To verify this hypothesis, this CL adds a DumpWithoutCrashing() call
if the hypothesis is violated.

Bug: 1310020
Change-Id: I070c696063a5c92506dd33a546055d9f9b078e74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3750920
Reviewed-by: vikas soni <vikassoni@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1022104}
NOKEYCHECK=True
GitOrigin-RevId: a6030cc74f9c8145bf424ebe89877bb68b4639c0