media/gpu/test: s/VideoDecoderClient/DecoderWrapper/ and cosmetic fixes
VideoDecoderClient is not a client at all, but a compound of the actual
VideoDecoder (usually a VideoDecoderPipeline) and associated logic. So
Wrapper feels like a better name.

Since I was already touching this, I did a few cosmetic changes:
- Member vars decoder_client_(thread_/sequence_checker_) are now called
"worker" which is a Chromium commonplace name.
- s/|video_decoder_client_state_|/|state_|/ since it's a clear internal
name.
- s/|video_player_sequence_checker_|/|parent_sequence_checker_|/ since
a class cannot be prescriptive about its owner class.
- Event response method are added an On... prefix (e.g. FlushDoneTask()
turns into OnFlushDoneTask(), which is clearer and more common in
Chromium.

Some comments are fixed.

Bug: None
Change-Id: Iea7395020f3e84febb6ec74d4ffbc6ebd030b179
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3751722
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1022243}
NOKEYCHECK=True
GitOrigin-RevId: b107a278fe61cf29c6fbc26bcdad517fa06cbcda