media/gpu/test: Simplify Thread and creation logic
DecoderWrapper owns a base::Thread which is discouraged [1]. Instead,
code should rely on a SingleThreadTaskRunner being created elsewhere
(this is to give Chrome flexibility to modify the thread pools in the
future). This CL does that.

Since such SingleThreadTaskRunner can't fail to be created, this has
the added side effect of greatly simplifying the construction logic
of the pair VideoPlayer+DecoderWrappper, eliminating a bit of code.

[1] https://source.chromium.org/chromium/chromium/src/+/main:base/threading/thread.h;l=31-33;drc=d9e4a00a8574aa6978e67d8d664db50e3f74edd4

Bug: None
Change-Id: Ib39cb78c6273cdd4726f8a516c8711965354be47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3754559
Auto-Submit: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023567}
NOKEYCHECK=True
GitOrigin-RevId: d9aef3ef69844313c01acddb618d6960e2ca319e