Wire-up active GPU luid to capture service on windows
A new mojo method is implemented and CaptureServiceProvider is listening to GpuDataManager updates now.
If the adapter is switched a message is sent to the video capture service.

Once DXGIDeviceManager is updated, IMFCaptureEngine picks up the new
device automatically.

Also, update ShMemBufferTracker to be reusable only for exactly the same pixel format as was used when creating the tracker. Otherwise, when switching LUID, one-off shmem tracker may be created (because IMFCaptureEngine is slow to pickup the new device). Then this tracker may be selected by the buffer pool to be used to put a texture in it, which won't work. The only way to distinguish that shmem tracker from gmb trackers is the pixel format.

Bug: 1336546
Change-Id: I564df4678ffdba3d634322d589eab63d5779548a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3720159
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Markus Handell <handellm@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Ilya Nikolaevskiy <ilnik@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024636}
NOKEYCHECK=True
GitOrigin-RevId: 12787307096f6c1e54c5048d7692b990a6508a4e