[Libcast-In-Chrome #2] Add new OpenscreenFrameSender
Patch:
This patch adds a new implementation of FrameSender, that wraps an
openscreen::cast::Sender in a new OpenscreenFrameSender class.

Conversion methods are included to allow interfacing with
openscreen::cast::Sender.

Finally, a new FrameSender::Create method that takes an
openscreen::cast::Sender is added, and types that consume FrameSender
(like AudioSender) have a new constructor added.

Coming up next:
1. Test coverage of the RemotingSender, AudioSender, and VideoSender
needs to be improved to allow testing without using a cast transport.

2. A new class (likely called OpenscreenSessionHost) needs to be
added to wrap an openscreen::cast::SenderSession.

3. Metrics need to be added to make sure we don't introduce major
regressions.

4. A new feature flag check needs to be added to instantiate an
OpenscreenSessionHost instead of the legacy Session type.

5. Once these pieces are in place, extensive manual testing and
improving of the automated test suite needs to occur.

??. Almost certainly some support classes and other changes will need
to be made.

6. Experimentation, bug fixing, and generally moving towards release!

Overall feature:
Libcast-In-Chrome is a project with the goal of having libcast be the
backing implementation of cast streaming in Chrome.

Bug: 1318499

Change-Id: I2a6504967b4b3b42c4990d10608416c21eb16313
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3708077
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025991}
NOKEYCHECK=True
GitOrigin-RevId: 4d1b8985684df03062b135b62a6b692a4fb97321