Revert "[CodecImage] Verify that CopyTexImage() is a no-op"
This CL reverts
https://chromium-review.googlesource.com/c/chromium/src/+/3750920,
which added a DumpWithoutCrashing() call to CodecImage::CopyTexImage().
We have now gathered crash reports from the invocation and it is
causing test failures. Hence we are taking it out, and we will re-add
it once we have reason to believe that we have solved the issues that
are causing it to go off.

Bug: 1310020, 1347077
Change-Id: I97eadde85ab33e0d3927c96e824e793e6868d397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3787834
Reviewed-by: vikas soni <vikassoni@chromium.org>
Commit-Queue: vikas soni <vikassoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1028242}
NOKEYCHECK=True
GitOrigin-RevId: a2ab0c5986204cb15400b0469143eddcdb99f7f0