Extract device selection from AudioManagerAndroid
This CL extracts and encapsulates logic from AudioManagerAndroid. It
moves audio devices monitoring and selection to AudioDeviceSelector,
without any changes in functionality. The AudioDeviceSelector can be
further subclassed based on the API level.

Some of the APIs we are currently using are deprecated on Android S and
above, and will be addressed in a follow-up CLs.

Some references to Android L and below are opportunistically removed,
now that we don't support it anymore.

Bug: 1317548
Change-Id: I809e3cdc01737c048f4bc0e1410384bd619b8ed6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3785461
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1028856}
NOKEYCHECK=True
GitOrigin-RevId: fceb4433d4eb4469aca0a337550e56b4e6267380