Generate a |release_token| for frames produced by VideoDecoderPipeline.
Prior to this CL, we only generated a |release_token| for VideoFrames
whose HasReleaseMailboxCB() returns true. This CL expands that to allow
the media::VideoDecoder implementation to specify whether it wants the
MojoVideoDecoderService to always register decoder frames for later
release.

The use case is out-of-process video decoding: in this case, the
media::VideoDecoder that backs the MojoVideoDecoderService in the video
decoder process outputs GpuMemoryBuffer VideoFrames without the
expectation of them being released with a SyncToken. However, they do
add a destruction observer to the VideoFrames so that the
media::VideoDecoder knows when it's okay to re-use the GpuMemoryBuffer.

The general flow of events is this:

1) The video decoder process outputs a GpuMemoryBuffer VideoFrame that
   gets to the GPU process.

2) The GPU process creates a SharedImage for this GpuMemoryBuffer and
   sends the corresponding Mailbox VideoFrame to the renderer process.

3) The renderer process uses the Mailbox VideoFrame in whatever way it
   wishes. It updates the release SyncToken for this frame. At some
   point, it releases the VideoFrame by calling
   media.mojom.VideoFrameHandleReleaser.ReleaseVideoFrame().

4) The GPU process waits on the release SyncToken of the
   MailboxVideoFrame before releasing the corresponding GpuMemoryBuffer
   VideoFrame using
   media.stable.mojom.VideoFrameHandleReleaser.ReleaseVideoFrame().

5) The MojoVideoDecoderService in the video decoder process releases the
   GpuMemoryBuffer VideoFrame thus activating its destruction observers
   which tell the underlying media::VideoDecoder that the
   GpuMemoryBuffer can be re-used.

Bug: b:195769334
Test: None
Change-Id: I7d4e2dfed07df3b49dd9f29d278314e2450ef08d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3741131
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1029390}
NOKEYCHECK=True
GitOrigin-RevId: 849d95d6e306ebe6f5d677c5bdca7d9f585ebb9a