Add H265VAAPIWrapper component for H.265 decoding.
This is commit 2/4 of a 4-commit CL that adds H.265 decoding support to the
decode-test binary. This commit adds the H265VAAPIWrapper component,
which is responsible for making calls to libva on behalf of the decoder.
This component is the only component in the decoding pipeline that
communicates with libva.

BUG=b:227247449

Change-Id: I7a9d91cd32a8ba696fde0c1793f09bddb03df986
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3778261
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Joey Allen <joeyallen@google.com>
Cr-Commit-Position: refs/heads/main@{#1033808}
NOKEYCHECK=True
GitOrigin-RevId: b82d8a4d82993b50973fb55280f8b6848865496a