Update decode_test platform decoding binary to support H.265
This is commit 4/4 of a 4-commit CL that adds H.265 decoding support to the
decode-test binary. This commit updates the decode-test binary to add
support for H265Decoding binary. This includes a few changes that allow
the H265Decoder to be selected from the command line.

BUG=b:227247449
TEST=tast run ${volteer|brya} video.PlatformDecoding.vaapi_hevc_main

Currently we are passing (120/135) test vectors. For a list of failing
test vectors please see b/239927523, b/239936640, b/239819547, and
b/239936226

Change-Id: I9e91d9419180aab5bef2c6eff2f00ed2fd159788
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3778263
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Joey Allen <joeyallen@google.com>
Cr-Commit-Position: refs/heads/main@{#1033833}
NOKEYCHECK=True
GitOrigin-RevId: d238b2c47085bd149b89e9e4035a7b21b3eca6e5