Don't cache context capabilities for pixel uploads
When we upload pixels inside SharedImage code, we need to make sure we
don't alter GL state. To do so we use ScopedResetAndRestoreUnpackState,
that queries related state and restores it after. Unfortunately, what
state needs to be restored depends on the GL context capabilities
(version/extensions).

Previously we queried capabilities in ctor of SharedImage factories
and cached it, which is not correct because upload can happen on a
different context than creation. This CL removes caching and does
query for each upload from the current context.

Change-Id: I94264d56049a05b6bd7092684c16dd80acd34891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3820361
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034139}
NOKEYCHECK=True
GitOrigin-RevId: 7c6d3bf5a70ba043c30d4a6e8d11099938a606dc