Fix Hang in MediaFoundationRendererTest::DirectCompositionHandle
There is a potential deadlock scenario in the DirectCompositionHandle
test when calling IMF* interface methods (in particular when we call
SetWindowlessSwapchain), where the underlying code attempts to acquire a
critical section but the critical section is blocked by another thread
attempting to call back into the thread originating the IMF* interface
call.

This is a result of the mock PMP server object being in a COM single
thread apartment (STA) instead of a multi threaded apartment (MTA).

This change just updates our COM initializer to use MTA to resolve this
issue.

Bug: 1319903
Change-Id: I91abf74f2cf7a85513f33e0261af75bed83ea37a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3826428
Reviewed-by: Frank Li <frankli@microsoft.com>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: William Carr <wicarr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1035159}
NOKEYCHECK=True
GitOrigin-RevId: 81ad161ff3f7c6538b02b9ce84f150c371d4a25d