HLS: Inform playlist parsing function of expected playlist version
This CL provides `hls::MediaPlaylist::Parse` and
`hls::MultivariantPlaylist::Parse` with the expected hls version of the
playlist being parsed. This is useful to know in advance since the rules
and semantics of various tags change depending on the version the
playlist declares, and the spec does not require the EXT-X-VERSION tag
in any specific part of the playlist. It does require that the version
does not change between playlist reloads, so the version can be
determined up front via `hls::Playlist::IdentifyPlaylist` and reused
across reloads.

Future CLs will plumb this version number down to the lower-level
parsing functions that require it. This CL adds basic tests to ensure
that the semantics of this parameter work as documented.

Bug: 1266991, 1284763
Change-Id: I747d5ca6c409159d4dda50a590f4e97bc2d5d166
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3823150
Commit-Queue: Will Cassella <cassew@chromium.org>
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1035763}
NOKEYCHECK=True
GitOrigin-RevId: 43aa4cd0d08a42c198653999b529b9f400c10e99