Video: Reassign the "configured" parameter sets with the latest "active" ones
Our previous logic uses active SPS/SPSext/PPS to be compared with the
configured SPS/SPSext/PPS and detect if it is not an IDR, if not,
then insert those active ones before the first slice picture.

There is a loophole in this logic. If a PPS appears in the middle of
the GOP, and it is a new one, our current logic is to insert it to the
VideoToolbox and this is correct, but if the next PPS with the same
PPS id before an IDR is just the same one that we've already configured with before the previous IDR picture, then we currently will consider configuration not changed, skip reconfigure decoder and ignore passing this PPS to Videotoolbox, and this will cause VideoToolbox uses the old PPS that in the middle of the GOP and cause issues.

(PS: I don't have a H264 like this, but i do found a H265 video sample that has a PPS in the middle of GOP (sample already uploaded in the related bug), and if i drag the timeline then it will cause a color shift.)

Reassign the "configured" PPS/SPS/SPSext with the latest "active" ones and trigger a reconfiguration at the next IDR picture to solve this bug is what this CL intend to do.

Bug: 1354448
Bug: 1306223
Change-Id: If9ab1e90b14e5098664cfd3dc4679a35fe86f324
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3840704
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1038294}
NOKEYCHECK=True
GitOrigin-RevId: 39a16b6dbf7e7ee278ba003278cd96b7f4536b4e