media: Refactor to store Key System Properties in a vector
In order for the experimental key system to be stored, this CL refactors
the code to use a vector to store key systems. Previously, the map could
not store two key systems with the same base key system name, as the map
stores the properties in a key by base key system names. Now, we store
the properties in a vector and retrieve the base names from the function
GetBaseKeySystemName() in KeySystemProperties.

Bug: b/241047837
Change-Id: I0ec5ab5c2c6be5473271cd9bf6e0ea8fd24f8458
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3852831
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1039021}
NOKEYCHECK=True
GitOrigin-RevId: 0452365a3d112e789ea5c6b6e0593a69344286e5