Reassign remaining non-specific Region Capture TODOs to cleanup
With the main work on Region Capture now done, and the feature shipped
as of m104, it is now time to close the tracking bug. Remaining TODOs
are minor cleanup, and are therefore assigned to the cleanup bug.

Bug: 1264849
Change-Id: Ie58f6d8532db9700d3aef934f02f775e26ec7ea1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3854088
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Auto-Submit: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1039434}
NOKEYCHECK=True
GitOrigin-RevId: 375e9085f62214e85e57b4fe591901dc1a188459