Add new AudioParameters constructors that are safe for DISCRETE layouts.
This CL chain aims to prevent a common bug where we create an
AudioParameters object with the ChannelLayout set to
CHANNEL_LAYOUT_DISCRETE, and then forget to call
set_channels_for_discrete(). This causes a crash when we try to stream
audio, because for DISCRETE the number of channels is defaulted to 0.

We prevent this by grouping the ChannelLayout and the number of channels
into a ChannelLayoutConfig, and using this in place of the ChannelLayout
when constructing AudioParameters. When constructing a ChannelLayoutConfig we either have to specify the number of channels, or
know at compile time that it's not DISCRETE.

Bug: 1286281
Change-Id: Iba47b406909d50c736ae432b30891909afe6ad13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3861967
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Fredrik Hernqvist <fhernqvist@google.com>
Cr-Commit-Position: refs/heads/main@{#1041450}
NOKEYCHECK=True
GitOrigin-RevId: 5b158e7c3a15b5131bf66d89df6db3d22ed140f2