vaapi: Enable AV1 protected content for AMD
This enables AV1 playback of protected content on AMD platforms that
support it. In the cdm-oemcrypto daemon, we take the transcrypted data
that AMD generates and hide it inside of a padding OBU. Then we extract
that here because that is the only data marked as the encrypted range.
We expect there to be 2 subsamples because there is a single trailing
unencrypted byte at the end of the padding OBU.

BUG=b:232908730
TEST=protected AV1 playback works on skyrim

Change-Id: Ie9843505ea8d49febd9cba9595a6c3851208f2e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696487
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/main@{#1042634}
NOKEYCHECK=True
GitOrigin-RevId: e547a08cbeb63a7d9cb457df8645d27cf666cab3