HLS: Make EXT-X-MEDIA tag DEFAULT=YES attribute imply AUTOSELECT=YES
My original interpretation of the spec was that if the AUTOSELECT
attribute is not present, it should *always* be considered "NO" even if
"DEFAULT=YES", despite the requirement that the only valid value in that
case is "YES". Given that this scenario is an implicit "YES", it's
useful to represent it that way when parsing the tag. This CL updates
the implementation and tests accordingly.

Bug: 1266991
Change-Id: I282997d264b7cbab96c54082a5b123b39b7a0a05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3859979
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1044740}
NOKEYCHECK=True
GitOrigin-RevId: f1cd38375b94bbedbfe098060f9ea19714dbde35