Separate player id from MediaLog.
In order to facilitate plumbing MediaLog into the browser process
while keeping it attached to the correct player, we need to remove
the ID from media_log itself. Logs are already tied together via
Mojo for IPC, so there isn't an issue of log records getting mixed
between players.

The player ID was really only needed for categorizing the events
on the frontends (media-internals and devtools), and the devtools
frontend already used a different unique ID.

So this CL removes the reliance on MediaLog for getting the logging
ID, and instead passes it where needed, including the renderers for
TRACE_EVENT macros, and to the render_media_event_handlers for
setting IDs for chrome://media-internals

R=liberato

Change-Id: I8574a4ba3dfb388294c79438b0fc62375c9abfd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3803211
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046273}
NOKEYCHECK=True
GitOrigin-RevId: 836aafa4b6ac7a1e5d05ec43990ee00bf06e4abf