media/gpu/v4l2VEA,IP: Workaround to USERPTR API against read-only buf
VideoFrame fed in VEA::Encode() is read-only since R107 if the
VideoFrame has a shared memory. V4L2VideoEncodeAccelerator fails
because VIDOC_QBUF with USERPTR buffer fails if the pointers
references a read-only buffer.
The USERPTR API issue is apparently to be fixed by a kernel side.
In the meantime, this CL adds the workaround to the issue; the read
only buffer is copied to a writable temporary buffer every Encode
before VIDIOC_QBUF.
Not that VideoFrame has a shared memory in the case of screen share
because a camera stack produces GpuMemoryBuffer.

Bug: b:243883312
Test: video_encode_accelerator_tests on elm, kevin and trogdor
Test: screen share in Google Meet on elm, kevin and trogdor
Change-Id: I922d98eaf52f80d9cd6c3784a8544bffb1232856
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3891421
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1046659}
NOKEYCHECK=True
GitOrigin-RevId: 203fdd03a4317f2eac57161ee101515a82df704f