[Libcast-in-Chrome #8] Introduce OpenscreenSessionHost
Patch:
This patch is really the core of the Libcast-in-Chrome migration,
adding a new OpenscreenSessionHost class that wraps an
openscreen::cast::SenderSession and launches it in the MirroringService
whenever the media::kOpenscreenCastStreamingSession flag is enabled.

Additional supporting changes to some mojoms, the MediaRemoter, and
test code are included.

For more context, see the reference patch that includes this plus other
changes:
https://chromium-review.googlesource.com/c/chromium/src/+/3868508.

Coming up next:
1. Cleanup and refactoring

2. An overview of metrics, logs, and tracing to determine where there
are opportunities for improvements.

3. Beginning release and working towards team dogfood.

Overall feature:
Libcast-In-Chrome is a project with the goal of having libcast be the
backing implementation of cast streaming in Chrome.

LOW_COVERAGE_REASON=can't use full GPU in tests. MediaRemoter coverage issue is tracked as https://crbug.com/1363265

Bug: 1318499, b/159352836
Change-Id: Icc8ce618842fbaa98f0ab8077fe5460f3e3565c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3873205
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047362}
NOKEYCHECK=True
GitOrigin-RevId: 4068f3275878d203c5f86623b3686673b989c0b6