media: Allow VdaVideoDecoder to work with OOP-VD.
This CL makes the changes needed to integrate the VdaVideoDecoder with
out-of-process video decoding.

The main problem we need to solve is to output GpuMemoryBuffer-backed
VideoFrames from the VdaVideoDecoder instead of Mailbox-backed
VideoFrames.

The general approach is to re-use the IMPORT mode of operation of the
ChromeOS VDAs (before this CL, this mode was only used for ARC++/ARCVM,
i.e., hardware video decoding for Android apps on ChromeOS). In this
mode, the VDA implementation is not expected to allocate its own buffers
for output and then bind those buffers to the textures created by the
PictureBufferManager. Instead, the VDA client is expected to allocate
the output buffers and then tell the VDA to import them by calling
ImportBufferForPicture(). Then, when the VDA calls PictureReady(), the
client can use its own reference to the output buffer to do whatever it
needs to do.

Implementation-wise, the most substantial changes are to the
PictureBufferManagerImpl: when operating in this IMPORT mode,
PictureBufferManagerImpl::CreatePictureBuffers() will allocate a
GpuMemoryBuffer-backed VideoFrame for each PictureBuffer. Then,
PictureBufferManagerImpl::CreateVideoFrame() will wrap this VideoFrame
into another instead of creating a Mailbox-backed VideoFrame (in IMPORT
mode, GL won't be involved at all).

Bug: b:195769334
Test: video.PlaybackPerf.h264_1080p_30fps_hw_oopvd on kevin
Change-Id: I74f29b6f100b701e0430e419ebc6eeaa1917e7f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3872253
Auto-Submit: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1047615}
NOKEYCHECK=True
GitOrigin-RevId: 35f0e9eb0cf7edab147664448421e78337510f82